Disable copy button on inline code
Change the hugo configuration for our template to display the
copy-to-clipboard button only on fenced or indented code blocks, not
inline code that is surrounded by single backticks.

In our docs, we are consistent about using fenced or indented code
blocks when we have an example that we intend someone to copy-paste, and
we use inline code to denote that something is literal.