Modernize ARO custom domain with cert docs (#327)
* Modernize ARO custom domain with cert docs

Switch instructions from using Fedora on WSL2 to using Bash in the Azure
Cloud Shell, to reduce the amount of setup someone needs to follow the
guide (however the guide should still be applicable in any unixy
environment). This also involves replacing the envsubst utility from
gettext with Bash interpolation, since gettext is annoying to install in
Cloud Shell.

Move instructions to fetch the console and API URLs until after zones
and domains have been set up to ensure that the reader doesn't try to
access their cluster's domains before they can resolve and think they've
done something wrong.

Add --scope argument to `az role assignment create` since it will be
required sometime in 2023 according to a deprecation message.

Clarify the wording of how to navigate the console in a couple of
places.

Make a couple of Markdown syntax fixes, as well as fix a couple of
capitalization and style typos.

* Fix pull secret instructions spelling & formatting

Correct the spelling of "Instead" and enclose the placeholder for the
secret in angle brackets to make it clearer that it's a placeholder.