Moving example of what ODF looks like to the end
The location where the example is displayed could lead the reader to the idea that the resource should already be available, while it is not. Only will be available after the step to create the storage cluster. So I'm moving this part to the end, after the creation of the storage cluster and the validation.