Revise ARO Quota instructions for new flow
Rather than submitting a support ticket, Azure now allows self-service
quota increase requests. The old support ticket flow now gives a button
which redirects to the new self-service flow.

Alter the instructions for "Adding Quota to ARO account" to account for
the new interface, and update the screenshot to match.

It would have been great to be able to do this all with "az quota"
commands on the CLI, but the CLI interface for quotas is still subject
to change.