Clarify wording around supported Arc features
Correct some potentially misleading language to make it clear what
Arc actually supports.

Replace "Azure Monitor" with "Container Insights" to make it clear
that we are talking about azuremonitor-containers and not
azuremonitor-metrics.

Replace "Observability" in heading with "Resource Viewing" to make
it clear that the section does not refer to metrics/logs/traces.

Fix a spelling error.