Remove unused ACM steps
There is a section that adds the cert to AWS ACM, and implies it will be
used by the NLB for TLS termination, but the NLB is never configured to
do so.

Since it's fine (arguably better) for the NLB to just handle things at
the TCP level and let the cluster handle TLS termination, remove this
section.