Reorder steps to put VM instructions together
Reorder things so that all of the AWS resources are created and
configured at once, and then all of the work is done on the VMs. This
will make it easier to follow the instructions once we have expanded to
doing everything VM-side on two VMs instead of just one.

Instructions remain the same, other than renumbering and a couple of
rewordings to still make the flow make sense.