Modify VPN instructions to be HA
Configure two IPsec VMs instead of just one. Use keepalived to ensure
that only one of them has libreswan running, and to assign a virtual IP
that other VMs can use as their next hop for VPC traffic.

Reorganize most on-VM steps into one section so that it's easy to repeat
for both VMs.

Switch to creating VMs using YAML since that allows us to set
anti-affinity for them.

For both ipsec and other VMs, configure the next-hop for the VPC traffic
route to be the virtual IP held by keepalived.

Add instructions to configure keepalived. We could have been fancier
about how we give it access to systemctl, but this should be good enough
on a box whose only purpose is the VPN. Also, we no longer have systemd
start the ipsec daemon since keepalived makes that decision instead.

Fix wording throughout the document to be consistent with the change to
a set of VMs.