Add HA for ipsec VMs (#667)
* Reorder steps to put VM instructions together

Reorder things so that all of the AWS resources are created and
configured at once, and then all of the work is done on the VMs. This
will make it easier to follow the instructions once we have expanded to
doing everything VM-side on two VMs instead of just one.

Instructions remain the same, other than renumbering and a couple of
rewordings to still make the flow make sense.

* Modify VPN instructions to be HA

Configure two IPsec VMs instead of just one. Use keepalived to ensure
that only one of them has libreswan running, and to assign a virtual IP
that other VMs can use as their next hop for VPC traffic.

Reorganize most on-VM steps into one section so that it's easy to repeat
for both VMs.

Switch to creating VMs using YAML since that allows us to set
anti-affinity for them.

For both ipsec and other VMs, configure the next-hop for the VPC traffic
route to be the virtual IP held by keepalived.

Add instructions to configure keepalived. We could have been fancier
about how we give it access to systemctl, but this should be good enough
on a box whose only purpose is the VPN. Also, we no longer have systemd
start the ipsec daemon since keepalived makes that decision instead.

Fix wording throughout the document to be consistent with the change to
a set of VMs.

* Move VPC route propagation to after IPsec VM setup

Reorder instructions to preserve the original ordering where propagating
route tables in the VPC and subnets happens after the VPN tunnel is up
to prevent traffic from being blackholed.

The most straighforward place to put this is after the ipsec VMs are
completely configured.

* Modify diagram to show HA

Add second VM and dotted lines to indicate the connections that would
take over upon failover.